feat: render parts of actions doc automatically from yaml
