+++ .github/workflows/auto-format.yml
@@ -40,9 +40,9 @@ jobs:
       - uses: quarto-dev/quarto-actions/setup@v2
         with:
           version: 1.4.515
-      - name: quarto render readme
+      - name: quarto render readme files
         run: |
-          quarto render README.qmd
+          quarto render **/README.qmd
       - name: format
         uses: pre-commit/action@v3.0.1
       - name: commit & push
